build: created internal connectins with render postgres

+++ config/contact_manager/views.py
@@ -1,5 +1,4 @@
 from django.shortcuts import render, redirect
-from django.test import Client
 from django.contrib.auth.decorators import login_required
 from django.views.generic.edit import CreateView
 from django.contrib.auth.forms import UserCreationForm
@@ -7,18 +6,15 @@ from django.urls import reverse_lazy
 import requests
 import json
 
-
 class RegisterView(CreateView):
     template_name = 'registration/register.html'
     form_class = UserCreationForm
     success_url = reverse_lazy('login')
 
-@login_required
 def display_contact_details(request, contact_id):
     api_url = f"http://localhost:8000/api/contacts/{contact_id}"    
     response = requests.get(api_url)
 
-    
     if response.status_code == 200:
         try:
             contact_data = response.json()
@@ -31,7 +27,6 @@ def display_contact_details(request, contact_id):
     else:
         error_message = f'Error fetching contact details: {response.status_code}'
         return render(request, 'contact_manager/error.html', {'error_message': error_message})
-
 
 @login_required
 def display_contact_list(request):
@@ -46,46 +41,12 @@ def display_contact_list(request):
     else:
         error_message = f'Error fetching contact list: {response.status_code}'
         return render(request, 'contact_manager/error.html', {'error_message': error_message})
-    
-
-@login_required
-def contact_edit(request, contact_id):
-    api_url = f"http://localhost:8000/api/contacts/{contact_id}"
-    
-    if request.method == 'POST':
-        # Handle form submission and update the contact
-        data = {
-            'first_name': request.POST['first_name'],
-            'last_name': request.POST['last_name'],
-            'phone': request.POST.get('phone', ''),
-            'email': request.POST['email'],
-            'address': request.POST.get('address', ''),
-            'city': request.POST.get('city', ''),
-            'state': request.POST.get('state', '')
-        }
-        response = requests.put(api_url, json=data)
-        
-        if response.status_code == 200:
-            return redirect('contacts')
-        else:
-            error_message = f'Error updating contact: {response.status_code}'
-            return render(request, 'contact_manager/error.html', {'error_message': error_message})
-    else:
-        response = requests.get(api_url)
-        
-        if response.status_code == 200:
-            contact_data = response.json()
-            return render(request, 'contact_manager/contact_edit.html', {'contact': contact_data})
-        else:
-            error_message = f'Error fetching contact details: {response.status_code}'
-            return render(request, 'contact_manager/error.html', {'error_message': error_message})
-
 
 @login_required
 def contact_delete(request, contact_id):
     api_url = f"http://localhost:8000/api/contacts/{contact_id}"
     
-    if request.method == 'POST':
+    if request.method == 'POST' and request.POST.get('_method') == 'DELETE':
         response = requests.delete(api_url)
         if response.status_code == 204:
             return redirect('contacts')
